refactor(recipes): unify detectSilentFail + captureForRunlog into stepObservation (#252)
Both modules wrap step output for the runner / runlog / dashboard but
operated as two independent files with no shared code or types. Merge
them into a single `src/recipes/stepObservation.ts` so the
"observability shims for step execution" concept lives in one place.

Behavior is unchanged. Both function signatures, all PATTERNS arrays,
the SENSITIVE_KEY_PATTERNS list, MAX_BYTES (8 KB), and the
`silentFailDetection: false` per-step opt-out all preserved verbatim.

- src/recipes/stepObservation.ts             — new merged module
- src/recipes/__tests__/stepObservation.test.ts — 33 tests merged from
  the two previous test files (22 + 11) with bodies untouched
- src/recipes/detectSilentFail.ts            — deleted
- src/recipes/captureForRunlog.ts            — deleted
- src/recipes/__tests__/detectSilentFail.test.ts — deleted
- src/recipes/__tests__/captureForRunlog.test.ts — deleted
- import sites updated:
    src/approvalHttp.ts
    src/recipes/chainedRunner.ts
    src/recipes/yamlRunner.ts

Closes #252.
